공유 상태에서 클릭하면 체크되게함. 검은화면 살짝 수정
