+++ src/components/Main/DayButton.js
@@ -10,7 +10,7 @@ const DayBtnSet = styled.div`
     position: absolute;
     background-color: white;
     width: 100%;
-    z-index: 1000;
+    z-index: 10;
     border-bottom: 1px solid white;
     padding-bottom: 26px;
 `
